jsonified response from server for chpater boards
