+++ server.py
@@ -1,7 +1,7 @@
 """Flask server for image differencing application"""
 from io import StringIO, BytesIO
 from datetime import datetime
-from flask import Flask, request, jsonify, render_template, flash, redirect, session
+from flask import Flask, request, json, jsonify, make_response, render_template, flash, redirect, session
 import os
 from PIL import Image
 import uuid
@@ -9,7 +9,7 @@ from werkzeug import secure_filename
 from werkzeug.datastructures import FileStorage
 
 from config import *
-from model import User, InputImage, DiffImage, ImageClass, UserClass
+from model import User, InputImage, DiffImage
 from utils import *
 
 
@@ -121,8 +121,20 @@ def main():
     username, user_id = current_user()
     user = UserClass(username)
     images = user.all_image_urls()
+    # chapters = {"board_00001": ['http://hackbright-image-upload-test.s3.amazonaws.com/guest/c55db769-59a9-470e-9678-0768c7b0d73e_static/images/guest_DODtrQ5W0AA-2S4.jpg', 'http://hackbright-image-upload-test.s3.amazonaws.com/guest/346b5e60-9dcc-43ec-9657-001cd0dbb49c_static/images/guest_IMG_5417.JPG'],
+    #             "board_00002": ['http://hackbright-image-upload-test.s3.amazonaws.com/guest/346b5e60-9dcc-43ec-9657-001cd0dbb49c_static/images/guest_IMG_5417.JPG', 'http://hackbright-image-upload-test.s3.amazonaws.com/guest/c7f56f81-8ebc-4533-8bd3-d0b605db2595_static/images/guest_IMG_5477.JPG', 'http://hackbright-image-upload-test.s3.amazonaws.com/guest/9d85f98c-f07b-4d6a-9b3c-cb7b697f4a13_static/images/guest_DODtrQ5W0AA-2S4.jpg', 'http://hackbright-image-upload-test.s3.amazonaws.com/guest/40b0e5bb-4bbc-4dae-85e6-436b8a6f64ee_static/images/guest_IMG_5417.JPG']}
 
-    return render_template("main.html", username=username, images=images)
+    chapters = {"boardName": "board_00001", "boardAssets": ['http://hackbright-image-upload-test.s3.amazonaws.com/guest/c55db769-59a9-470e-9678-0768c7b0d73e_static/images/guest_DODtrQ5W0AA-2S4.jpg', 'http://hackbright-image-upload-test.s3.amazonaws.com/guest/346b5e60-9dcc-43ec-9657-001cd0dbb49c_static/images/guest_IMG_5417.JPG']}, {"boardName": "board_00002", "boardAssets": ['http://hackbright-image-upload-test.s3.amazonaws.com/guest/346b5e60-9dcc-43ec-9657-001cd0dbb49c_static/images/guest_IMG_5417.JPG', 'http://hackbright-image-upload-test.s3.amazonaws.com/guest/c7f56f81-8ebc-4533-8bd3-d0b605db2595_static/images/guest_IMG_5477.JPG', 'http://hackbright-image-upload-test.s3.amazonaws.com/guest/9d85f98c-f07b-4d6a-9b3c-cb7b697f4a13_static/images/guest_DODtrQ5W0AA-2S4.jpg', 'http://hackbright-image-upload-test.s3.amazonaws.com/guest/40b0e5bb-4bbc-4dae-85e6-436b8a6f64ee_static/images/guest_IMG_5417.JPG']}
+    chapters = json.dumps(chapters)
+    chapters = json.loads(chapters)
+    # chapters = jsonify(chapters)
+    print ("IMAGES:", images)
+    print ("CHAPTERS:", chapters)
+
+    return render_template("main.html",
+                            username=username,
+                            images=images,
+                            chapters=chapters)
 
 
 # @app.route("/submit-diff-request", methods=['POST'])
